more refactoring to use new html methods etc

+++ src/wbs/console/misc/PageBuilder.java
@@ -92,7 +92,7 @@ class PageBuilder {
 	}
 
 	public
-	int pages () {
+	long pages () {
 		return pages.size ();
 	}
